issue #1: created basic tab interface and empty screens for future issues
