+++ frontend/android/settings.gradle
@@ -1,4 +1,6 @@
 rootProject.name = 'RentEasy'
+include ':react-native-vector-icons'
+project(':react-native-vector-icons').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-vector-icons/android')
 include ':react-native-maps'
 project(':react-native-maps').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-maps/lib/android')
 include ':@react-native-community_geolocation'
