+++ frontend/ios/Podfile
@@ -31,6 +31,8 @@ target 'RentEasy' do
 
   pod 'react-native-maps', :path => '../node_modules/react-native-maps'
 
+  pod 'RNVectorIcons', :path => '../node_modules/react-native-vector-icons'
+
   target 'RentEasyTests' do
     inherit! :search_paths
     # Pods for testing
