+++ frontend/package.json
@@ -11,7 +11,12 @@
     "@react-native-community/geolocation": "^2.0.2",
     "react": "16.8.6",
     "react-native": "0.60.5",
-    "react-native-maps": "^0.25.0"
+    "react-native-gesture-handler": "^1.4.1",
+    "react-native-maps": "^0.25.0",
+    "react-native-paper": "^2.16.0",
+    "react-native-vector-icons": "^6.6.0",
+    "react-navigation": "^4.0.10",
+    "react-navigation-material-bottom-tabs": "^2.1.2"
   },
   "devDependencies": {
     "@babel/core": "7.6.0",
