+++ frontend/screens/HomeScreenMap.js
@@ -1,0 +1,101 @@
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ * @flow
+ */
+
+import React, {Fragment} from 'react';
+import {
+  SafeAreaView,
+  StyleSheet,
+  ScrollView,
+  View,
+  Text,
+  StatusBar,
+} from 'react-native';
+
+import {
+  Colors,
+} from 'react-native/Libraries/NewAppScreen';
+
+import Geolocation from '@react-native-community/geolocation';
+import FetchLocation from '../components/FetchLocation';
+import UsersMap from '../components/UsersMap'
+import tabBarIcon from '../components/tabBarIcon';
+
+export default class HomeScreenMap extends React.Component {
+  static navigationOptions = {
+    tabBarIcon: tabBarIcon('md-map'),
+  };
+
+  state = {
+    userLocation: null
+  }
+
+  getUserLocationHandler = () => {
+    Geolocation.getCurrentPosition(position => {
+      this.setState({
+        userLocation: {
+          latitude: position.coords.latitude,
+          longitude: position.coords.longitude,
+          latitudeDelta: 0.0622,
+          longitudeDelta: 0.0421,
+        }
+      });
+    }, err => console.log(err));
+  }
+
+  render () {
+    return (
+          <ScrollView
+            contentInsetAdjustmentBehavior="automatic"
+            style={styles.scrollView}>
+            <View style={styles.container}>
+              <FetchLocation onGetLocation={this.getUserLocationHandler} />
+              <UsersMap userLocation={this.state.userLocation}/>
+            </View>
+          </ScrollView>
+    );
+  };
+};
+
+const styles = StyleSheet.create({
+  scrollView: {
+    backgroundColor: Colors.lighter,
+  },
+  engine: {
+    position: 'absolute',
+    right: 0,
+  },
+  body: {
+    backgroundColor: Colors.white,
+  },
+  sectionContainer: {
+    marginTop: 32,
+    paddingHorizontal: 24,
+  },
+  sectionTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+    color: Colors.black,
+  },
+  sectionDescription: {
+    marginTop: 8,
+    fontSize: 18,
+    fontWeight: '400',
+    color: Colors.dark,
+  },
+  highlight: {
+    fontWeight: '700',
+  },
+  footer: {
+    color: Colors.dark,
+    fontSize: 12,
+    fontWeight: '600',
+    padding: 4,
+    paddingRight: 12,
+    textAlign: 'right',
+  },
+});
